backend/requirements.txt
flask
flask-sqlalchemy
flask-login
flask-cors
requests
python-dotenv
openai==0.28 